# apps/api-gateway/openapi.yaml
openapi: 3.0.3
info:
  title: DONTAI api-gateway
  version: 0.1.0
  description: >
    Auth + routing layer in front of the DONTAI stack (PROJECT_DESIGN.md
    §2/§3-1). See README.md for scope notes -- this file documents the
    HTTP contract only, generated by hand from the actual Express routes
    (src/routes/*.ts), not the other way around; if this ever drifts from
    the code, the code is correct.
servers:
  - url: http://localhost:4000
    description: local dev
  - url: http://Philosophyz.iptime.org:4000
    description: Pi deployment

tags:
  - name: auth
  - name: profile
  - name: artworks

paths:
  /health:
    get:
      summary: Liveness check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /auth/signup:
    post:
      tags: [auth]
      summary: Create an account (email + password) and provision a custodial wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error (zod)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "409":
          description: Email or handle already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"

  /auth/login:
    post:
      tags: [auth]
      summary: Exchange email + password for a token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"

  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Refresh token invalid, expired, or its user no longer exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"

  /auth/{provider}:
    get:
      tags: [auth]
      summary: Start the OAuth authorization-code flow for a social provider
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, kakao]
      responses:
        "302":
          description: Redirect to the provider's own authorize page, with a CSRF state param
        "501":
          description: Provider has no <PROVIDER>_CLIENT_ID/SECRET configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"

  /auth/{provider}/callback:
    get:
      tags: [auth]
      summary: >
        OAuth callback. Exchanges the code for a profile, finds or creates the
        user (matched on authProvider+providerUserId), and redirects to
        `${WEB_URL}/oauth-callback#accessToken=...&refreshToken=...` (tokens
        in the URL fragment so they never reach server access logs).
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, kakao]
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to WEB_URL/oauth-callback with tokens in the URL hash
        "400":
          description: Missing/invalid/expired state (CSRF check failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "502":
          description: Token exchange with the provider failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"

  /me:
    get:
      tags: [profile]
      summary: Profile of the bearer token's owner
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: User profile (passwordHash/encryptedWalletKey never included)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Missing/invalid/expired bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: Token valid but the user row no longer exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"

  /artworks:
    post:
      tags: [artworks]
      summary: >
        Create an artwork. creatorId and ownerWalletAddress come from the
        verified JWT, never from the request body -- proxies straight to
        asset-service's POST /artworks with those injected.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtworkRequest"
      responses:
        "202":
          description: Upload accepted, protection pipeline running in the background
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: ast_2029e10078884a87
                  status:
                    type: string
                    example: UPLOADED
        "400":
          description: Validation error
        "401":
          description: Missing/invalid bearer token
        "502":
          description: asset-service unreachable
    get:
      tags: [artworks]
      summary: List the caller's own artworks (proxies asset-service's GET /artworks?creatorId=)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Artwork list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artwork"
        "401":
          description: Missing/invalid bearer token
        "502":
          description: asset-service unreachable

  /artworks/{id}:
    get:
      tags: [artworks]
      summary: Get one artwork's full detail (proxies asset-service's GET /artworks/:id)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artwork detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "401":
          description: Missing/invalid bearer token
        "404":
          description: No such artwork
        "502":
          description: asset-service unreachable

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SignupRequest:
      type: object
      required: [email, password, handle]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        handle:
          type: string
          minLength: 3
          maxLength: 32
          pattern: "^[a-zA-Z0-9_]+$"
        displayName:
          type: string

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserSummary"

    UserSummary:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        handle:
          type: string
        walletAddress:
          type: string
        role:
          type: string
          enum: [USER, CREATOR, MODERATOR, ADMIN]

    UserProfile:
      allOf:
        - $ref: "#/components/schemas/UserSummary"
        - type: object
          properties:
            displayName:
              type: string
              nullable: true
            avatarUri:
              type: string
              nullable: true
            createdAt:
              type: string
              format: date-time
            status:
              type: string

    CreateArtworkRequest:
      type: object
      required: [title, sourceImageUri]
      properties:
        title:
          type: string
        sourceImageUri:
          type: string
          description: >
            Local file path in this PoC, not a real object-storage URI
            (see apps/asset-service/README.md).
        protectionProfile:
          type: string
          enum: [L1_PREVIEW, L2_PORTFOLIO, L3_ANTI_TRAIN]
          default: L3_ANTI_TRAIN
        allowAiTraining:
          type: boolean
          default: false

    Artwork:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        sourceImageUri:
          type: string
        creatorId:
          type: string
        ownerWalletAddress:
          type: string
        protectionProfile:
          type: string
          enum: [L1_PREVIEW, L2_PORTFOLIO, L3_ANTI_TRAIN]
        allowAiTraining:
          type: boolean
        status:
          type: string
          enum: [UPLOADED, PROTECTING, REGISTERING, PUBLISHED, FAILED]
        errorMessage:
          type: string
          nullable: true
        protectedImageUri:
          type: string
          nullable: true
        perceptualHash:
          type: string
          nullable: true
        metadataHash:
          type: string
          nullable: true
        watermarkPayloadHex:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        assetVersions:
          type: array
          items:
            type: object
        ownershipRecords:
          type: array
          items:
            type: object
            properties:
              ownerWallet:
                type: string
              contentHash:
                type: string
              chain:
                type: string
              registryAddress:
                type: string
              txHash:
                type: string
              blockNumber:
                type: integer
              registeredAt:
                type: string
                format: date-time

    ValidationError:
      type: object
      properties:
        error:
          type: object
          description: zod's .flatten() output

    ErrorMessage:
      type: object
      properties:
        error:
          type: string